# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  handler.CreateReviewReq:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      reviewee_id:
        type: integer
      trade_request_id:
        type: integer
    required:
    - rating
    - reviewee_id
    - trade_request_id
    type: object
...
  handler.SendMessageReq:
    properties:
      content:
        type: string
      conversation_id:
        type: integer
      message_type:
        type: string
      receiver_id:
        type: integer
    required:
    - content
    type: object
...
  model.ConversationParticipant:
    properties:
      conversation_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_read_at:
        type: string
      unread_count:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: integer
      created_at:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      message_type:
        type: string
      sender:
        $ref: '#/definitions/model.User'
      sender_id:
        type: integer
    type: object
  model.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      reviewee:
        $ref: '#/definitions/model.User'
      reviewee_id:
        type: integer
      reviewer:
        $ref: '#/definitions/model.User'
      reviewer_id:
        type: integer
      trade_request_id:
        type: integer
    type: object
...
  model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      credit_score:
        type: integer
      enterprise_name:
        type: string
      enterprise_verified:
        type: boolean
      id:
        type: integer
      last_login_at:
        type: string
      location:
        type: string
      nickname:
        type: string
      phone:
        type: string
      real_name_verified:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: NeoBarter Team
    url: https://github.com/sxc19960907/neobarter
  description: 现代易货交易平台后端 API 文档。支持 C2C/B2B 双模式、巴特币结算、智能匹配。
  license:
    name: MIT
  termsOfService: https://github.com/sxc19960907/neobarter
  title: 新易巴特 NeoBarter API
  version: "1.0"
paths:
...
  /messages:
    post:
      consumes:
      - application/json
      description: 指定 conversation_id 或 receiver_id（自动创建会话）
      parameters:
      - description: 消息内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SendMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Message'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发送消息
      tags:
      - 消息
  /messages/conversations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ConversationParticipant'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取会话列表
      tags:
      - 消息
  /messages/conversations/{id}:
    get:
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取会话消息
      tags:
      - 消息
  /messages/conversations/{id}/read:
    put:
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 标记会话已读
      tags:
      - 消息
  /notifications:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取通知列表
      tags:
      - 通知
  /notifications/{id}/read:
    put:
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 标记单条通知已读
      tags:
      - 通知
  /notifications/read-all:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 标记全部通知已读
      tags:
      - 通知
  /notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取未读通知数
      tags:
      - 通知
  /reviews:
    post:
      consumes:
      - application/json
      description: 交易完成后对对方评分（1-5星），影响信用积分
      parameters:
      - description: 评价内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 提交交易评价
      tags:
      - 评价
  /reviews/user/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户评价列表
      tags:
      - 评价
  /search/items:
    get:
      description: 全文搜索 + 多条件筛选，返回高亮结果
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      - description: 成色
        in: query
        name: condition
        type: string
      - description: 最低估值
        in: query
        name: min_value
        type: number
      - description: 最高估值
        in: query
        name: max_value
        type: number
      - description: 地区
        in: query
        name: location
        type: string
      - description: 排序
        in: query
        name: sort_by
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索物品（Elasticsearch）
      tags:
      - 搜索
  /search/suggest:
    get:
      parameters:
      - description: 前缀关键词
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 搜索自动补全
      tags:
      - 搜索
...
securityDefinitions:
  BearerAuth:
    description: 输入格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"